src/main/java/com/acoustic/service/AnnualNetService.java
package com.acoustic.service;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Component;

import java.math.BigDecimal;
import java.math.RoundingMode;

@Component
@RequiredArgsConstructor
public class AnnualNetService implements SalaryCalculatorService{

    public static final int MONTHS_NUMBER = 12;

    @Override
    public String getDescription() {
        return "Annual net";
    }

    @Override
    public BigDecimal apply(final BigDecimal grossMonthlySalary) {
        return grossMonthlySalary.multiply(BigDecimal.valueOf(MONTHS_NUMBER)).setScale(2, RoundingMode.HALF_EVEN);
    }
}
